Fix broken navigation on sm screen (for slide-over menu):
* Remove '/inventory' layout.svelte
* Include the "layout" in both '/inventory/inbound/+page.svelte' and '/inventory/warehouses/+page.svelte'
* TODO: We might want to reorganise this (maybe create an 'InventoryPage' component)
